[#61] Concrete Bezier curve features
